+++ packages/egg-layer/test/fixtures/eaas-fc/app/router.js
@@ -1,0 +1,13 @@
+'use strict';
+
+/**
+ * @param {Egg.Application} app - egg application
+ */
+module.exports = app => {
+  const { router, controller } = app;
+  router.get('/get', controller.home.getMethod);
+  router.get('/get/query', controller.home.getQueryMethod);
+  router.post('/post', controller.home.postMethod);
+  router.post('/post/body', controller.home.postBodyMethod);
+  router.get('/buffer', controller.home.buffer);
+};