+++ packages/serverless-spec-builder/src/wrapper.ts
@@ -31,6 +31,10 @@ export function writeWrapper(options: {
     middleware,
   } = options;
   const files = {};
+  if (service?.deployType) {
+    expandYMLForApplication(service);
+  }
+
   const functions = service.functions || {};
   for (const func in functions) {
     const handlerConf = functions[func];
@@ -64,7 +68,15 @@ export function writeWrapper(options: {
       });
     }
   }
-  const tpl = readFileSync(resolve(__dirname, '../wrapper.ejs')).toString();
+
+  const isCustomAppType = !!service?.deployType;
+
+  const tpl = readFileSync(
+    resolve(
+      __dirname,
+      isCustomAppType ? '../wrapper_app.ejs' : '../wrapper.ejs'
+    )
+  ).toString();
   // for function programing，function info build to dist/faasFunctionsMap.json
   // need register this function to ioc
   const functionMapFile = resolve(distDir, faasFunctionsMapFile);
@@ -81,7 +93,7 @@ export function writeWrapper(options: {
       copyFileSync(sourceFile, registerFunctionFile)
     }
   }
-  
+
   for (const file in files) {
     const fileName = join(distDir, `${file}.js`);
     const layers = getLayers(service.layers, ...files[file].originLayers);
@@ -121,4 +133,40 @@ export function formetAggregationHandlers(handlers) {
       }
       return handlerB.level - handlerA.level;
     });
+}
+
+/**
+ * support application deploy to serverless cloud platform
+ * @param service
+ */
+function expandYMLForApplication(service) {
+  if (!service.deployType) {
+    return;
+  }
+
+  // add default function
+  if (!service.functions) {
+    service.functions = {
+      app_index: {
+        handler: 'index.handler',
+        events: [{ http: { path: '/*' } }],
+      },
+    };
+  }
+
+  if (service?.layers) {
+    service.layers = {};
+  }
+
+  if (service?.deployType === 'egg') {
+    service.layers['eggLayer'] = { path: 'npm:@midwayjs/egg-layer' };
+  }
+
+  if (service?.deployType === 'express') {
+    service.layers['expressLayer'] = { path: 'npm:@midwayjs/express-layer' };
+  }
+
+  if (service?.deployType === 'koa') {
+    service.layers['koaLayer'] = { path: 'npm:@midwayjs/koa-layer' };
+  }
 }